+++ index.js
@@ -46,5 +46,5 @@ module.exports = {
         return false;
     }
     return true;
-  }
+  },
 };